+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# Unreleased
+
+Update to use react-native-web (instead of deprecated WebView from react-native package)
+
 # 1.0.0
 
 - [BREAKING] Updated default style of the webview from `{ width: '100%', height: '100%' }` to `{ flex: 1 }`